# cheaper collectFields cache key