## What it Does

This makes the internal `collectFields` function cheaper. For small selections, it allocates more, but is faster. For large selections, it allocates significantly less and is much, much faster.

Before:

```
% go test -v ./graphql/executor -bench=BenchmarkExecuteRequest -run asdqweqwe -benchtime 5s
goos: darwin
goarch: arm64
pkg: github.com/ccbrown/api-fu/graphql/executor
BenchmarkExecuteRequest
BenchmarkExecuteRequest-10    	    6174	    879445 ns/op	  796542 B/op	   20444 allocs/op
```

After:

```
% go test -v ./graphql/executor -bench=BenchmarkExecuteRequest -run asdqweqwe -benchtime 5s
goos: darwin
goarch: arm64
pkg: github.com/ccbrown/api-fu/graphql/executor
BenchmarkExecuteRequest
BenchmarkExecuteRequest-10    	    6068	    846561 ns/op	  828877 B/op	   22444 allocs/op
```

## Steps to Test

<!-- All changes should have automated tests when feasible: -->

`go test -v ./...` and see above benchmark commands.

<!-- Does running this require any special setup or dependencies? -->